#5450 refactoring, renaming
https://github.com/metasfresh/metasfresh/issues/5450
