+++ de.metas.adempiere.adempiere/migration/src/main/sql/postgresql/system/10-de.metas.adempiere/5529430_sys_gh5450_RenameProcess_MutateCustomerPrices.sql
@@ -1,0 +1,25 @@
+-- 2019-08-27T10:23:45.756Z
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Process SET Classname='de.metas.pricing.process.M_PricelistVersion_MutateCustomerPrices', Name='Preislisten mutieren',Updated=TO_TIMESTAMP('2019-08-27 13:23:45','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Process_ID=541180
+;
+
+-- 2019-08-27T10:23:59.882Z
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Process_Trl SET IsTranslated='Y', Name='Mutate Customer Prices',Updated=TO_TIMESTAMP('2019-08-27 13:23:59','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Language='en_US' AND AD_Process_ID=541180
+;
+
+-- 2019-08-27T10:24:29.756Z
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Process_Trl SET IsTranslated='Y', Name='Preislisten mutieren',Updated=TO_TIMESTAMP('2019-08-27 13:24:29','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Language='de_DE' AND AD_Process_ID=541180
+;
+
+-- 2019-08-27T10:24:35.997Z
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Process_Trl SET IsTranslated='Y', Name='Preislisten mutieren',Updated=TO_TIMESTAMP('2019-08-27 13:24:35','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Language='de_CH' AND AD_Process_ID=541180
+;
+
+-- 2019-08-27T10:24:44.668Z
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Process_Trl SET Name='Preislisten mutieren',Updated=TO_TIMESTAMP('2019-08-27 13:24:44','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Language='nl_NL' AND AD_Process_ID=541180
+;
+
