+++ de.metas.materialtracking/src/main/java/de/metas/materialtracking/qualityBasedInvoicing/impl/MaterialTrackingDocumentsPricingInfo.java
@@ -13,7 +13,6 @@ import java.util.TreeMap;
 import org.adempiere.ad.dao.ICompositeQueryFilter;
 import org.adempiere.ad.dao.IQueryBL;
 import org.adempiere.ad.dao.impl.CompareQueryFilter.Operator;
-import de.metas.location.CountryId;
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.adempiere.util.lang.ImmutablePair;
 import org.adempiere.util.lang.ObjectUtils;
@@ -30,6 +29,7 @@ import com.google.common.collect.ListMultimap;
 
 import de.metas.document.engine.IDocument;
 import de.metas.lang.SOTrx;
+import de.metas.location.CountryId;
 import de.metas.materialtracking.IMaterialTrackingPPOrderBL;
 import de.metas.materialtracking.model.IMaterialTrackingAware;
 import de.metas.materialtracking.model.I_M_InOutLine;
@@ -288,7 +288,7 @@ import de.metas.util.Services;
 					.addInArrayOrAllFilter(I_M_InOut.COLUMN_DocStatus, IDocument.STATUS_Completed, IDocument.STATUS_Closed)
 					.addCompareFilter(I_M_InOut.COLUMN_MovementDate, Operator.GREATER_OR_EQUAL, plv.getValidFrom());
 
-			final I_M_PriceList_Version nextPLV = priceListDAO.retrieveNextVersionOrNull(plv);
+			final I_M_PriceList_Version nextPLV = priceListDAO.retrieveNextVersionOrNull(plv, true /* onlyProcessed */ );
 			if (nextPLV != null)
 			{
 				inOutFilter.addCompareFilter(I_M_InOut.COLUMN_MovementDate, Operator.LESS, nextPLV.getValidFrom());
